invoices and categories

+++ saba/admin/pages/groups/[...].tsx
@@ -1,22 +1,28 @@
+import { useLazyQuery, useMutation } from '@apollo/client';
+import { PageContainer } from '@keystone-6/core/admin-ui/components';
+import { Button } from "@keystone-ui/button";
+import { Stack, useTheme } from '@keystone-ui/core';
+import { FieldContainer, FieldLabel, TextArea, TextInput } from "@keystone-ui/fields";
+import { ChevronRightIcon, ClipboardIcon, MoreVerticalIcon, PlusIcon } from '@keystone-ui/icons';
+import { LoadingDots } from '@keystone-ui/loading';
 import {
     Alert,
     Box,
     Divider, Drawer, Grid, List, ListItem, ListItemButton, ListItemIcon, ListItemSecondaryAction, ListItemText,
+    Modal,
     Paper,
     Snackbar,
-} from '@mui/material'
-import MuiButton from '@mui/material/Button'
-import React from 'react'
-import { ActivityIcon, PlusIcon, MoreVerticalIcon, ChevronRightIcon } from '@keystone-ui/icons';
-import { gql } from '@ts-gql/tag/no-transform';
-import { useLazyQuery, useMutation } from '@apollo/client';
-import { useRouter } from 'next/router';
+    Tab,
+    Tabs,
+    Typography,
+} from '@mui/material';
+import MuiButton from '@mui/material/Button';
 import { createTheme, styled, ThemeProvider } from '@mui/material/styles';
-import { LoadingDots } from '@keystone-ui/loading'
-import { FieldContainer, FieldLabel, TextInput, TextArea } from "@keystone-ui/fields";
-import { Stack } from '@keystone-ui/core'
-import { PageContainer } from '@keystone-6/core/admin-ui/components';
-import { Button, ButtonContext } from "@keystone-ui/button"
+import { gql } from '@ts-gql/tag/no-transform';
+import { useRouter } from 'next/router';
+import React from 'react';
+import { useDebouncedValue } from '../../../data/utils';
+
 
 type PageState = {
     snackbarIsOpen: boolean,
@@ -47,6 +53,153 @@ const GroupIcon: React.FC<{ width: string, height: string }> = ({ width: w, heig
     return <svg xmlns="http://www.w3.org/2000/svg" width={w} height={h} viewBox="0 0 24 24"><g fill="none" stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5}><circle cx={17} cy={7} r={3}></circle><circle cx={7} cy={17} r={3}></circle><path d="M14 14h6v5a1 1 0 0 1-1 1h-4a1 1 0 0 1-1-1zM4 4h6v5a1 1 0 0 1-1 1H5a1 1 0 0 1-1-1z"></path></g></svg>
 }
 
+function ChangeParent(props: { value: string | null, id: string, onChange: (i: string) => void }) {
+
+
+    const [isOpen, setIsOpen] = React.useState(false)
+    const [newParentId, setNewParentId] = React.useState<string>('')
+    const debounceNewParent = useDebouncedValue(newParentId, 1300)
+    const theme = useTheme()
+
+    const CATEGORY_BY_ID = gql`
+    query CATEGORY_BY_ID($id: ID!) {
+        category(where:  { id:  $id} ) {
+            id
+            title
+            code
+        }
+    }` as import("../../../__generated__/ts-gql/CATEGORY_BY_ID").type
+
+    const UPDATE_CATEGORY = gql`
+mutation UPDATE_CATEGORY($id: ID!, $data: CategoryUpdateInput!) {
+    updateCategory(where: { id: $id }, data: $data) {
+        id
+        title
+        code
+    }
+}
+` as import("../../../__generated__/ts-gql/UPDATE_CATEGORY").type
+
+
+    const [load, { data, loading, refetch }] = useLazyQuery(CATEGORY_BY_ID, {})
+    const [updateCategoryMutation, { loading: updateLoading }] = useMutation(UPDATE_CATEGORY)
+    const [pageState, setPageState] = React.useContext(GroupContext)
+
+    React.useEffect(() => {
+        if (debounceNewParent)
+            load({ variables: { id: debounceNewParent }, fetchPolicy: "no-cache" })
+    }, [debounceNewParent])
+
+    const isLoading = React.useMemo(() => {
+        return loading || updateLoading || newParentId !== debounceNewParent
+    }, [loading, newParentId, debounceNewParent, updateLoading])
+
+    async function tryUpdate() {
+        if (data?.category) {
+            try {
+
+                const res = await updateCategoryMutation({
+                    variables: {
+                        id: props.id,
+                        data: {
+                            parent: { connect: { id: debounceNewParent } }
+                        }
+                    }
+                })
+
+                if (res.errors)
+                    throw new Error(res.errors[0].message)
+
+                props.onChange(data.category.id)
+                setTimeout(() => {
+                    window.location.reload()
+                }, 1000);
+            } catch (error) {
+                setPageState({ ...pageState, snackbarIsOpen: true, snackbarMessage: String(error), snackbarSeverity: "error" })
+            }
+            setIsOpen(false)
+        }
+    }
+
+    const style = {
+        position: 'absolute' as 'absolute',
+        top: '50%',
+        left: '50%',
+        transform: 'translate(-50%, -50%)',
+        width: 400,
+        bgcolor: 'background.paper',
+        border: '2px solid #ccc',
+        boxShadow: 24,
+        p: 4,
+        borderRadius: "8px"
+    };
+
+    return <>
+
+        <MuiButton color="primary" variant="contained" onClick={() => { setIsOpen(true) }}>
+            نغییر دسته ی پدر
+        </MuiButton>
+
+        <Modal
+            open={isOpen}
+            onClose={() => {
+                setIsOpen(false)
+            }}
+            aria-labelledby="modal-modal-title"
+            aria-describedby="modal-modal-description"
+        >
+            <Box sx={style}>
+                <Typography id="modal-modal-title" variant="h5" component="h2">
+                    تغییر دسته ی والد
+                </Typography>
+                <Typography id="modal-modal-description" sx={{ my: 2 }}>
+                    این دسته بندی را به زیرمجموعه ی دسته ی دیگری
+                    منتقل کنید
+                </Typography>
+
+                <FieldContainer>
+                    <FieldLabel>
+                        شناسه
+                        والد جدید
+
+                    </FieldLabel>
+                    <TextInput
+
+                        placeholder={props.value || ''}
+                        value={newParentId}
+                        onChange={e => {
+                            setNewParentId(e.target.value)
+                        }}
+                    />
+                    <div style={{ height: "6px" }} >
+                        {
+                            isLoading && <LoadingDots size='small' label='loading' />
+                        }
+                        {
+                            !isLoading && debounceNewParent !== '' && !data?.category?.id && <span style={{ padding: theme.spacing.small, color: theme.tones.negative.fill['1'] }}>
+                                not found
+                            </span>
+                        }
+                        {
+
+                            !isLoading && data?.category?.id && <span style={{ color: theme.tones.positive.fill['1'] }}>
+                                {data.category.title} - {data.category.code}
+                            </span>
+                        }
+                    </div>
+                </FieldContainer>
+                <Box sx={{ display: "flex", justifyContent: "flex-end", mt: 2 }} >
+
+                    <Button isLoading={updateLoading} tone='active' onClick={async () => { tryUpdate() }}  >
+                        به روز رسانی
+                    </Button>
+                </Box>
+            </Box>
+        </Modal>
+
+    </>
+}
+
 const Group: React.FC<{
     parentId: string | null,
     index: number,
@@ -71,7 +224,6 @@ const Group: React.FC<{
 
     const router = useRouter();
     const myParam = router.query[""]?.at(props.index + 1) || '' as string
-    console.log({ myParam })
 
     const [open, setOpen] = React.useState(false);
     const [selectedItemToEdit, setSelectedItemToEdit] = React.useState<null | string>("");
@@ -89,6 +241,8 @@ const Group: React.FC<{
     }, [])
 
 
+
+
     if (loading) return <div style={{ padding: "20px" }} ><LoadingDots label='loading' tone='passive' /></div>
 
     return (
@@ -102,6 +256,7 @@ const Group: React.FC<{
                         refetch()
                     }}
                 />
+
             </Drawer>
             <List dense >
                 {
@@ -169,6 +324,7 @@ const CategoryEditor: React.FC<CategoryEditorProps> = (props) => {
     })
 
     const [pageState, setPageState] = React.useContext(GroupContext);
+    const [tabIndex, setTabIndex] = React.useState(0);
 
     const CATEGORY_DETAIL = gql`
     query CATEGORY_DETAIL($id: ID!) {
@@ -232,6 +388,13 @@ mutation DELETE_CATEGORY($id: ID!) {
         })
     }, [])
 
+    function a11yProps(index: number) {
+        return {
+            id: `simple-tab-${index}`,
+            'aria-controls': `simple-tabpanel-${index}`,
+        };
+    }
+
     if (loading)
         return (
             <Box sx={{ width: 350 }} padding={2} role="presentation">
@@ -240,35 +403,49 @@ mutation DELETE_CATEGORY($id: ID!) {
         )
 
     return (
-        <Box sx={{ width: 350 }} height={'100%'} padding={2} role="presentation">
-            <FieldContainer style={{ display: 'flex', flexDirection: 'column', height: '100%' }} >
-                <div style={{ marginTop: "20px" }}>
-                    <FieldLabel>title</FieldLabel>
-                    <Stack gap="small">
+        <>
+            <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
+                <Tabs value={tabIndex} onChange={(_, nv) => {
+                    setTabIndex(nv)
+                }} aria-label="basic tabs example">
+                    <Tab label={props.itemId === null ? "ایجاد گروه" : "ویرایش گروه"} {...a11yProps(0)} />
+                    <Tab label="تنظیمات پیشرفته" {...a11yProps(1)}
+                        disabled={props.itemId === null}
+                    />
+                </Tabs>
+            </Box>
+            {/* tab panel */}
+            <Box sx={{ width: 450, display: 'flex', flexDirection: 'column' }} height={'100%'} padding={2} role="presentation">
 
-                        <TextInput value={editableInfo.title} onChange={e => setEditableInfo({ ...editableInfo, title: e.target.value })} />
-                    </Stack>
-                </div>
-                <div style={{ marginTop: "20px" }}>
+                <div style={{ display: tabIndex === 0 ? 'block' : 'none' }} >
 
-                    <FieldLabel  >code</FieldLabel>
-                    <Stack gap="small">
+                    <FieldContainer style={{ display: 'flex', flexDirection: 'column', height: '100%' }} >
+                        <div style={{ marginTop: "20px" }}>
+                            <FieldLabel>title</FieldLabel>
+                            <Stack gap="small">
 
-                        <TextInput value={editableInfo.code} onChange={e => setEditableInfo({ ...editableInfo, code: e.target.value })} />
-                    </Stack>
-                </div>
-                <div style={{ marginTop: "20px" }}>
+                                <TextInput value={editableInfo.title} onChange={e => setEditableInfo({ ...editableInfo, title: e.target.value })} />
+                            </Stack>
+                        </div>
+                        <div style={{ marginTop: "20px" }}>
 
-                    <FieldLabel  >Description</FieldLabel>
-                    <Stack gap="small">
+                            <FieldLabel  >code</FieldLabel>
+                            <Stack gap="small">
 
-                        <TextArea value={editableInfo.description} onChange={e => setEditableInfo({ ...editableInfo, description: e.target.value })} />
-                    </Stack>
-                </div>
+                                <TextInput value={editableInfo.code} onChange={e => setEditableInfo({ ...editableInfo, code: e.target.value })} />
+                            </Stack>
+                        </div>
+                        <div style={{ marginTop: "20px" }}>
 
-                <Grid container spacing={2} style={{ marginTop: "auto" }}>
-                    <Grid item xs={props.itemId ? 8 : 12}>
-                        <Button tone='active' weight='bold' style={{ width: "100%" }} onClick={async () => {
+                            <FieldLabel  >Description</FieldLabel>
+                            <Stack gap="small">
+
+                                <TextArea value={editableInfo.description} onChange={e => setEditableInfo({ ...editableInfo, description: e.target.value })} />
+                            </Stack>
+                        </div>
+
+
+                        <Button tone='active' weight='bold' style={{ width: "100%", marginTop: "20px" }} onClick={async () => {
 
                             try {
 
@@ -337,29 +514,61 @@ mutation DELETE_CATEGORY($id: ID!) {
 
                             {props.itemId ? "ویرایش" : "افزودن"}
                         </Button>
-                    </Grid>
-                    {props.itemId && <Grid item xs={4} >
-                        <Button tone='negative' style={{ width: "100%" }}
-                            onClick={async () => {
-                                //delete
-                                await deleteCategory({ variables: { id: props.itemId! } })
 
-                                props.onSubmit()
-                                // refetch()
-                                props.onClose()
-                            }
 
-                            } >
-                            حذف
-                        </Button>
-                    </Grid>
-                    }
-                </Grid>
+                    </FieldContainer>
+                </div>
+                {props.itemId && <>
+                    <div style={{ display: tabIndex === 1 ? 'block' : 'none' }} >
+                        <FieldContainer style={{ display: 'flex', flexDirection: 'column', height: '100%' }} >
+                            <FieldLabel>id</FieldLabel>
+                            <div style={{ display: 'flex', gap: 4 }} >
 
-            </FieldContainer>
-        </Box>
+                                <TextInput value={props.itemId} disabled />
+                                <Button style={{ display: 'flex', gap: 4, justifyContent: 'center', alignItems: 'center' }}
+                                    onClick={() => {
+                                        // TODO IMPEL ADD TO CLIPBOARD!!
+                                        setPageState({
+                                            ...pageState,
+                                            snackbarIsOpen: true,
+                                            snackbarSeverity: 'success',
+                                            snackbarMessage: 'در کلیپبرد ذخیره شد'
+                                        })
+                                    }}
+                                >
+                                    <ClipboardIcon size={16} />
+                                </Button>
+                            </div>
+                            <Divider textAlign="center" sx={{ marginTop: 4, marginBottom: 1 }} > Danger Zone</Divider>
+                            <Stack gap='large' >
+
+                                <FieldLabel>change parent</FieldLabel>
+                                <ChangeParent id={props.itemId} value={props.parentId} onChange={i => console.log(i)} />
+
+                                <MuiButton variant="outlined" color="error"
+                                    onClick={async () => {
+                                        if (!window.confirm("آیا از حذف اطمینان دارید؟"))
+                                            return
+
+                                        //delete
+                                        await deleteCategory({ variables: { id: props.itemId! } })
+
+                                        props.onSubmit()
+                                        // refetch()
+                                        props.onClose()
+                                    }}
+                                >
+                                    <span>حذف</span>
+                                </MuiButton>
+                            </Stack>
+                        </FieldContainer>
+                    </div></>}
+            </Box>
+        </>
+
     );
 }
+
 
 
 export default function GroupsPage() {
